% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacity.R
\name{capacity}
\alias{capacity}
\title{Calculate Projected Battery Capacity}
\usage{
capacity(csv)
}
\arguments{
\item{csv}{A data frame containing the necessary columns for `kWhUsed()`
and `percentUsed()`, which are used to calculate energy consumption and
battery percentage used.}
}
\value{
The projected full battery capacity in kilowatt-hours (kWh).
}
\description{
This function estimates the total battery capacity in kilowatt-hours (kWh)
based on energy consumption and the percentage of battery used during a
logging session.
}
\examples{
\dontrun{
data <- data.frame(SoC = c(90, 80, 70), Distance = c(0, 10, 20))
capacity(data)  # Returns projected capacity based on
    kWhUsed(data) and percentUsed(data)
}
}
\seealso{
kWhUsed, percentUsed
}
