% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chargingCurve.R
\name{chargingCurve}
\alias{chargingCurve}
\title{Generate a Charging Curve}
\usage{
chargingCurve(csv)
}
\arguments{
\item{csv}{A data frame containing a log output recorded while DC fast
charging the vehicle.}
}
\value{
A line plot of charging power (kW) over time.
}
\description{
This function uses ggplot2 to create a charging curve using an OBD log of
a DC fast charging session. (x: Time, y: Power (kW))
}
\examples{
\dontrun{
plot <- chargingCurve(evCalc::log_charging)
}
}
