% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degradation.R
\name{degradation}
\alias{degradation}
\title{Calculate Battery Degradation}
\usage{
degradation(csv, kWhNew)
}
\arguments{
\item{csv}{A data frame containing the necessary columns for `kWhUsed()`
and `percentUsed()`, which are used to calculate energy consumption and
battery percentage used.}

\item{kWhNew}{A numeric value representing the original full battery
capacity (in kWh) when the battery was new.}
}
\value{
The percentage decrease in usable battery capacity.
}
\description{
This function computes the percentage decrease in usable battery capacity
compared to its original capacity when new.
}
\examples{
\dontrun{
data <- data.frame(SoC = c(90, 80, 70), Distance = c(0, 10, 20))
degradation(data, kWhNew = 75)  # Returns battery degradation percentage
}
}
\seealso{
kWhUsed, percentUsed, capacity
}
